requirements.txt
asn1crypto==1.5.1
pypdf==4.0.1
python_dateutil==2.8.2
streamlit==1.31.1
